Refactor daemon subpackage
