+++ daemon/src/gke_cluster.ts
@@ -1,32 +1,13 @@
-import * as protosTypes from "@google-cloud/container/build/protos/protos"
+import { KubectlClient, GkeClient, ClusterConfig, getName } from "./types"
 
-import { V1PodList, V1NamespaceList } from "@kubernetes/client-node";
 const DEFAULT_NODE_POOL_NAME = "default-pool"
-
-interface GkeClient {
-    getProjectId(): Promise<string>;
-    setNodePoolSize(request: protosTypes.google.container.v1.ISetNodePoolSizeRequest): Promise<[protosTypes.google.container.v1.IOperation, protosTypes.google.container.v1.ISetNodePoolSizeRequest | undefined, {} | undefined]>;
-}
-interface KubectlClient {
-    listNamespace(): Promise<{body: V1NamespaceList}>;
-    listNamespacedPod(namespace: string): Promise<{body: V1PodList}>;
-}
-
-export interface ClusterConfig {
-    clusterName: string;
-    zone: string;
-    project?: string;
-}
 
 export class GkeCluster {
     constructor(private config: ClusterConfig, private gkeClient: GkeClient,
                 private kubectlClient: KubectlClient) {
     }
     private async getDefaultNodePoolName(): Promise<string> {
-        if (this.config.project == null) {
-            this.config.project = await this.gkeClient.getProjectId()
-        }
-        return `projects/${this.config.project}/zones/${this.config.zone}/clusters/${this.config.clusterName}/nodePools/${DEFAULT_NODE_POOL_NAME}`
+        return `${await getName(this.config, this.gkeClient)}/nodePools/${DEFAULT_NODE_POOL_NAME}`
     }
 
     public async activate(): Promise<void> {
